+++ src/Data/AddressQueryData.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace SmartDato\GlsItaly\Data;
+
+use Spatie\LaravelData\Data;
+
+final class AddressQueryData extends Data
+{
+    public function __construct(
+        public string $province,
+        public string $zipcode,
+        public string $city,
+        public string $street,
+    ) {}
+}